Add extra module file for interactive.js
